Combine test_util.py and test_util_fcns.py
