+++ .circleci/config.yml
@@ -51,4 +51,4 @@ jobs:
             source activate myenv
             conda activate RAiDER
             cd test
-            py.test test_interpolator.py test_util_fcns.py test_hdf5_parallel.py
+            py.test test_interpolator.py test_util.py test_hdf5_parallel.py
